Move station components around
